13- idealstar(nf,id,,MOD). Exceptions may occur in ideallog [#2549]
In particular when the "generator" for a cyclic F_q^* (\otimes Z/MOD)
doesn't have order q-1 but some strict divisor. In this case, the first
stage of the computation of the discrete log (reducing to input = 1 (mod I))
may fail and we get a meaningless err_COPRIME exception.

Solution: store MOD in the structure in idealstar, and use it in ideallog