move QXQ_charpoly [#2591] fix to its proper place
QXQ_charpoly arguments are documented as QX + ZX and this is easy
and cheaper to ensure in the caller.

Move fix to RgXQ_charpoly_fast with slight inefficency (extra
gerepileupto) in the silly case where the T argument was *not* a ZX.
Cheaper than a (systematic!) Q_primpart in any case, and a priori never
exercised if the caller of RgXQ_charpoly is sane.

I would go so far as calling RgXQ_charpoly with a non-ZX argument T as a
bug if it occurs through a libpari call (OK through a random call to
GP's charpoly: we can't prevent users from being inefficient).