heuristic initial precision increase in bnfinit to help small_snorm [#2632]
Commit 784f1ae504 removed this precision estimate
1) it had become possibly wrong when Fincke_Pohst_bound was introduced
2) it was used to reduce the precision of M to a "sufficient" value,
   potentially causing problems for a minor efficiency gain

Before that commit, the bound was valid to compute norms reliably from
floating point embeddings and nothing else. But as a side effect, it
could increase the initial nfinit precision if a t_POL was input. Which
would help in other places, such as small_norm's gaussred_from_QR which
is breaking here.

1) is not a serious concern, the value is of the right order of magnitude
and this is a one-sided error: at worse we would fail to compute largish
vectors which are unlikely to be smooth anyway.
2) can be circumvented: never reduce the precision, but possibly let the
   bound increase it.

So reinstate the bound as a better heuristic guess for the initial
precision. It won't hurt, it may help.