31- elliptic functions ignored ellperiods(,1) quasi-period component
This is the description for the 2nd and following hunks of commit
b4cb096a06 which were meant as two successive commits but were unfortunately
committed all at once.

This commits also fixes a nasty problem when w1/w2 is on the border of the
fundamental domain, breaking ellperiod([w1,w2]) idempotence. Wich
becomes a problem when the input quasi-periods are taken into
consideration since red_modSL2 could change the lattice basis.